+++ test/s3/lifecycle/s3_lifecycle_versioning_test.go
@@ -7,6 +7,7 @@ package lifecycle
 import (
 	"context"
 	"fmt"
+	"strconv"
 	"strings"
 	"testing"
 	"time"
@@ -15,6 +16,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/seaweedfs/seaweedfs/weed/pb/filer_pb"
+	"github.com/seaweedfs/seaweedfs/weed/s3api/s3_constants"
 	"github.com/stretchr/testify/require"
 )
 
@@ -54,6 +56,16 @@ func putNoncurrentExpirationLifecycle(t *testing.T, c *s3.Client, bucket, prefix
 
 // backdateVersionedMtime ages a specific .versions/v_<id> entry. The
 // version files live under <buckets>/<bucket>/<key>.versions/v_<versionId>.
+//
+// If the entry already carries an ExtNoncurrentSinceNsKey stamp (written
+// by the S3 PUT handler at demotion time), the stamp is rewritten to the
+// same backdated instant so the lifecycle engine's noncurrent clock —
+// which prefers the explicit stamp over the legacy sibling-mtime
+// derivation — also moves backward. Without this the tests would see a
+// real-now stamp dominate a backdated mtime and the rule would never
+// fire. A test simplification: production stamps record when the
+// successor was written, not the demoted version's own mtime; the test
+// only needs the resulting clock to be old enough.
 func backdateVersionedMtime(t *testing.T, fc filer_pb.SeaweedFilerClient, bucket, key, versionID string, daysOld int) {
 	t.Helper()
 	dir := bucketsPath + "/" + bucket + "/" + key + ".versions"
@@ -65,8 +77,14 @@ func backdateVersionedMtime(t *testing.T, fc filer_pb.SeaweedFilerClient, bucket
 	require.NotNil(t, resp.Entry)
 	require.NotNil(t, resp.Entry.Attributes)
 
-	resp.Entry.Attributes.Mtime = time.Now().Add(-time.Duration(daysOld) * 24 * time.Hour).Unix()
+	backdated := time.Now().Add(-time.Duration(daysOld) * 24 * time.Hour)
+	resp.Entry.Attributes.Mtime = backdated.Unix()
 	resp.Entry.Attributes.MtimeNs = 0
+	if _, hasStamp := resp.Entry.Extended[s3_constants.ExtNoncurrentSinceNsKey]; hasStamp {
+		resp.Entry.Extended[s3_constants.ExtNoncurrentSinceNsKey] = []byte(
+			strconv.FormatInt(backdated.UnixNano(), 10),
+		)
+	}
 	_, err = fc.UpdateEntry(context.Background(), &filer_pb.UpdateEntryRequest{
 		Directory: dir, Entry: resp.Entry,
 	})
